feat: replace all interfaces with types
